Allow traffic from knative-eventing and knative-serving
Workflows like move2kube deploy and use knative functions to achieve
the workflow tasks and working with functions means we need to allow the
traffic from knative-eventing and knative-serving namespace.

Traffic from knative-serving is needed when the function is deployed in
the same workflow namespace, because the way the function is triggered is
by the knative-serving components that send the event to the function.
So the traffic is knative-serving -> $workflows-ns (where the knative
function is deployed)

Traffic from knative-eventing is needed when functions or in the future
other events go through the knative broker and the trigger is sending
them to workflows. In the case of move2kube it is the
m2k-save-transformation function that sends an event, and that event
direction is knative-eventing -> $workflows-ns

Signed-off-by: Roy Golan <rgolan@redhat.com>